refactor: IBoardList 인터페이스 board.type.ts 파일로 분리하여 재사용성 개선
