+++ trip-talk/src/app/components/BoardList/BoardList.tsx
@@ -6,29 +6,22 @@ import {
 } from "../../../commons/graphql/backend-api";
 import { useRouter } from "next/navigation";
 import { useMutation } from "@apollo/client";
-
-interface IBoardList {
-  id: string;
-  number: number;
-  writer: string;
-  title: string;
-  createdAt: string;
-}
+import { IBoardList } from "../../../types/board.type";
 
 export default function BoardList(props: IBoardList) {
-  const { id, number, title, writer, createdAt } = props;
+  const { _id, number, title, writer, createdAt } = props;
   const router = useRouter();
   const [deleteBoard] = useMutation(DELETE_BOARD);
 
   const opnClickBoard = () => {
-    router.push(`/routes/boards/${id}`);
+    router.push(`/routes/boards/${_id}`);
   };
 
   const onCLickDelete = (event: React.MouseEvent) => {
     event.stopPropagation();
     deleteBoard({
       variables: {
-        id: id,
+        id: _id,
       },
       refetchQueries: [{ query: FETCH_BOARDS }],
     });
