+++ trip-talk/src/types/board.type.ts
@@ -1,0 +1,7 @@
+export interface IBoardList {
+  _id: string;
+  number: number;
+  writer: string;
+  title: string;
+  createdAt: string;
+}
